requirements.txt
Flask
torch
clip
Pillow
numpy
SQLAlchemy
psycopg2-binary
